+++ web-server/api/src/main/java/com/hotplace/api/ApiApplication.java
@@ -2,7 +2,9 @@ package com.hotplace.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class ApiApplication {
