--- web-server/api/src/main/java/com/hotplace/api/dong/controller/DongDto.java
@@ -1,16 +1,0 @@
-package com.hotplace.api.dong.controller;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class DongDto {
-    private int dongId;
-    private String dongName;
-
-    public DongDto(int dongId, String dongName){
-        this.dongId = dongId;
-        this.dongName = dongName;
-    }
-}
