+++ web-server/api/src/main/java/com/hotplace/api/dto/GuDto.java
@@ -1,0 +1,16 @@
+package com.hotplace.api.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class GuDto {
+    private int guId;
+    private String guName;
+
+    public GuDto(int guId, String guName){
+        this.guId = guId;
+        this.guName = guName;
+    }
+}
