+++ web-server/api/src/main/java/com/hotplace/api/entity/Category.java
@@ -1,0 +1,17 @@
+package com.hotplace.api.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@NoArgsConstructor
+@Table (name = "CATEGORY")
+@Entity
+public class Category {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "category_id")
+    private Integer id;
+    private String name;
+}
