--- web-server/api/src/main/java/com/hotplace/api/gu/domain/Gu.java
@@ -1,21 +1,0 @@
-package com.hotplace.api.gu.domain;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-@Table (name = "GU")
-@Entity
-public class Gu {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "gu_id")
-    private Integer id;
-
-    @Column(name = "gu_name")
-    private String name;
-
-}
