--- web-server/api/src/main/java/com/hotplace/api/gu/domain/GuRepository.java
@@ -1,9 +1,0 @@
-package com.hotplace.api.gu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface GuRepository extends JpaRepository<Gu, Integer> {
-    List<Gu> findAll();
-}
