--- web-server/api/src/main/java/com/hotplace/api/place/controller/IndexController.java
@@ -1,18 +1,0 @@
-package com.hotplace.api.place.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class IndexController {
-
-    @GetMapping("/index")
-    public String index() {
-        return "index.html";
-    }
-
-    @GetMapping("")
-    public String place(){
-        return "place.html";
-    }
-}
