--- web-server/api/src/main/java/com/hotplace/api/place/controller/PlacePostController.java
@@ -1,36 +1,0 @@
-package com.hotplace.api.place.controller;
-
-import com.hotplace.api.api_form.ApiForm;
-import com.hotplace.api.place.dto.PlaceRequest;
-import com.hotplace.api.place.dto.PlaceResponse;
-import com.hotplace.api.place.service.PlaceService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api")
-public class PlacePostController {
-
-    private final PlaceService placeService;
-
-    @GetMapping("/places")
-    public ApiForm<List<PlaceResponse>> findAllPlacesByDong(@RequestParam("dong") Integer id){
-        System.out.println("test :"+id);
-        return placeService.findAllByDong(id);
-    }
-
-    @PostMapping("/places")
-    public ApiForm<List<PlaceResponse>> searchPlaces(@RequestBody PlaceRequest requestDto){
-        return placeService.searchPlaceRequest(requestDto);
-    }
-
-    @PostMapping("/paging/places")
-    public ApiForm<Page<PlaceResponse>> pagePlaces(@RequestBody PlaceRequest requestDto, Pageable pageable){
-        return placeService.searchPage(requestDto, pageable);
-    }
-}
