+++ web-server/api/src/main/java/com/hotplace/api/repository/DongRepository.java
@@ -1,0 +1,10 @@
+package com.hotplace.api.repository;
+
+import com.hotplace.api.entity.Dong;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface DongRepository extends JpaRepository<Dong, Integer> {
+    List<Dong> findAllByGuId(Integer id);
+}
